+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Add explicit table name to `Consumer` to prevent implicit lookup through pluralization.
+- Use compiled GraphQL schema by default (glesys/butler-graphql:^12.1 required).
 
 
 ## [0.29.0] - 2025-08-14